requirements.txt
flask
opencv-python
opencv-python-headless
numpy
Pillow
